refactor: [MON-310] LikeToggleWrapped를 SeriesLikeToggle로 수정
