+++ src/components/mocules/LikeToggle/SeriesLikeToggle.jsx
@@ -1,12 +1,10 @@
-// onClick에 api호출 함수 담아서 전달  await addLikeSeries(id); await delLikeSeries(id);
-// userInfo 호출해서 로그인 유무 값 전달
 import React from 'react';
 import PropTypes from 'prop-types';
 import { addLikeSeries, delLikeSeries } from '@apis/like';
 import { useUser } from '@contexts/UserProvider';
-import { Like } from '@mocules';
+import { LikeToggle } from '@mocules';
 
-export const LikeWrapped = ({ id, isLiked, initialCount }) => {
+export const SeriesLikeToggle = ({ id, isLiked, initialCount }) => {
   const { userInfo } = useUser();
 
   const handleClick = async state => {
@@ -14,7 +12,7 @@ export const LikeWrapped = ({ id, isLiked, initialCount }) => {
   };
 
   return (
-    <Like
+    <LikeToggle
       seriesId={id}
       isLogin={userInfo.userId}
       isLiked={isLiked}
@@ -24,15 +22,15 @@ export const LikeWrapped = ({ id, isLiked, initialCount }) => {
   );
 };
 
-LikeWrapped.defaultProps = {
+SeriesLikeToggle.defaultProps = {
   id: 1,
   isLiked: false,
   initialCount: false,
 };
-LikeWrapped.propTypes = {
+SeriesLikeToggle.propTypes = {
   id: PropTypes.number,
   isLiked: PropTypes.bool,
   initialCount: PropTypes.oneOfType([PropTypes.number, PropTypes.bool]),
 };
 
-export default LikeWrapped;
+export default SeriesLikeToggle;
